stonk-ponk-be/requirements.txt
Django==3.1.7
mongo==0.2.0
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
pandas
yahoo-fin==0.8.7
googlenews==1.5.7
arrow
